Revert splineflags change causing players to have walk toggled on
Also added uint32 to enum so that this is not needed in the first place.

+++ src/game/Movement/packet_builder.cpp
@@ -67,7 +67,7 @@ namespace Movement
         // add fake Enter_Cycle flag - needed for client-side cyclic movement (client will erase first spline vertex after first cycle done)
         splineflags.enter_cycle = move_spline.isCyclic();
         // add fake Runmode flag - client has strange issues without that flag
-        data << uint32(splineflags & static_cast<uint32>(~MoveSplineFlag::Mask_No_Monster_Move | MoveSplineFlag::Runmode));
+        data << uint32((splineflags & ~MoveSplineFlag::Mask_No_Monster_Move) | MoveSplineFlag::Runmode);
         data << move_spline.Duration();
     }
